requirements.txt
pytube3==9.6.4
selenium==3.141.0
typing-extensions==3.7.4.2
urllib3==1.25.9
